skipping test in junctiontree tests
